Fix issue (3) in #28 - FileTreeview.ordered_insert now handles files and folders correctly
